Intcode refactor to use longs complete, old tests passing, ready for day 10

+++ Runner/Day07.cs
@@ -97,7 +97,7 @@ namespace Runner
                     Intcode intcode = intcodes[i];
                     intcode.InputQueue.Enqueue(lastOutput);
                     intcode.Resume();
-                    lastOutput = intcode.OutputQueue.Dequeue();
+                    lastOutput = (int)intcode.OutputQueue.Dequeue();
                 }
             } while (feedback && !intcodes[phases.Length - 1].Halt);
             return lastOutput;
